requirements-dev.txt
# test
pytest==8.3.2
syrupy==4.7.1
moto==5.0.13

# boto3
boto3==1.34.145
botocore==1.34.145
boto3-stubs[essential]==1.34.145
requests==2.32.3

# cdk
aws-cdk-lib==2.154.1
constructs>=10.0.0,<11.0.0
cdk-nag==2.28.189

# powertools
aws-lambda-powertools[aws-sdk]==2.22.0
